server: Separate creator vs. user management of license keys
