+++ server/polar/user/auth.py
@@ -3,7 +3,7 @@ from typing import Annotated
 from fastapi import Depends
 
 from polar.auth.dependencies import Authenticator
-from polar.auth.models import Anonymous, AuthSubject, Organization, User
+from polar.auth.models import Anonymous, AuthSubject, User
 from polar.auth.scope import Scope
 
 _UserBenefitsRead = Authenticator(
@@ -63,33 +63,6 @@ UserDownloadablesRead = Annotated[
                 Scope.user_downloadables_read,
             },
             allowed_subjects={User},
-        )
-    ),
-]
-
-LicenseKeysRead = Annotated[
-    AuthSubject[User | Organization],
-    Depends(
-        Authenticator(
-            required_scopes={
-                Scope.web_default,
-                Scope.user_license_keys_read,
-                Scope.user_license_keys_write,
-            },
-            allowed_subjects={User, Organization},
-        )
-    ),
-]
-
-LicenseKeysWrite = Annotated[
-    AuthSubject[User | Organization],
-    Depends(
-        Authenticator(
-            required_scopes={
-                Scope.web_default,
-                Scope.user_license_keys_write,
-            },
-            allowed_subjects={User, Organization},
         )
     ),
 ]
