+++ server/tests/user/endpoints/test_license_keys.py
@@ -8,9 +8,9 @@ from polar.benefit.schemas import (
     BenefitLicenseKeysCreateProperties,
 )
 from polar.kit.utils import generate_uuid, utc_now
+from polar.license_key.service import license_key as license_key_service
 from polar.models import Organization, Product, User
 from polar.postgres import AsyncSession
-from polar.user.service.license_key import license_key as license_key_service
 from tests.fixtures.auth import AuthSubjectFixture
 from tests.fixtures.database import SaveFixture
 from tests.fixtures.license_key import TestLicenseKey
@@ -18,134 +18,7 @@ from tests.fixtures.license_key import TestLicenseKey
 
 @pytest.mark.asyncio
 @pytest.mark.http_auto_expunge
-class TestLicenseKeyCRUDEndpoints:
-    @pytest.mark.auth(
-        AuthSubjectFixture(subject="user"),
-        AuthSubjectFixture(subject="organization"),
-    )
-    async def test_get_non_existing_404s(
-        self,
-        session: AsyncSession,
-        client: AsyncClient,
-    ) -> None:
-        random_id = generate_uuid()
-        response = await client.get(f"/v1/users/license-keys/{random_id}")
-        assert response.status_code == 404
-
-    async def test_get_unauthorized_401(
-        self,
-        session: AsyncSession,
-        client: AsyncClient,
-        save_fixture: SaveFixture,
-        user: User,
-        organization: Organization,
-        product: Product,
-    ) -> None:
-        benefit, granted = await TestLicenseKey.create_benefit_and_grant(
-            session,
-            save_fixture,
-            user=user,
-            organization=organization,
-            product=product,
-            properties=BenefitLicenseKeysCreateProperties(
-                prefix="testing",
-                expires=None,
-                limit_activations=None,
-            ),
-        )
-        id = granted["license_key_id"]
-        lk = await license_key_service.get(session, id)
-        assert lk
-
-        response = await client.get(f"/v1/users/license-keys/{id}")
-        assert response.status_code == 401
-
-    @pytest.mark.auth(
-        AuthSubjectFixture(subject="user"),
-        AuthSubjectFixture(subject="organization"),
-    )
-    async def test_get_authorized(
-        self,
-        session: AsyncSession,
-        client: AsyncClient,
-        save_fixture: SaveFixture,
-        user: User,
-        organization: Organization,
-        product: Product,
-    ) -> None:
-        benefit, granted = await TestLicenseKey.create_benefit_and_grant(
-            session,
-            save_fixture,
-            user=user,
-            organization=organization,
-            product=product,
-            properties=BenefitLicenseKeysCreateProperties(
-                prefix="testing",
-                expires=None,
-                limit_activations=None,
-            ),
-        )
-        id = granted["license_key_id"]
-        lk = await license_key_service.get(session, id)
-        assert lk
-
-        response = await client.get(f"/v1/users/license-keys/{id}")
-        assert response.status_code == 200
-        data = response.json()
-        assert data.get("benefit_id") == str(benefit.id)
-        assert data.get("key").startswith("TESTING")
-
-    @pytest.mark.auth(
-        AuthSubjectFixture(subject="organization"),
-    )
-    async def test_update(
-        self,
-        session: AsyncSession,
-        client: AsyncClient,
-        save_fixture: SaveFixture,
-        user: User,
-        organization: Organization,
-        product: Product,
-    ) -> None:
-        benefit, granted = await TestLicenseKey.create_benefit_and_grant(
-            session,
-            save_fixture,
-            user=user,
-            organization=organization,
-            product=product,
-            properties=BenefitLicenseKeysCreateProperties(
-                prefix="testing",
-                expires=None,
-                limit_activations=None,
-                limit_usage=None,
-            ),
-        )
-        id = granted["license_key_id"]
-        lk = await license_key_service.get(session, id)
-        assert lk
-
-        expires_at = utc_now() + relativedelta(months=1)
-        expires_at = expires_at.strftime("%Y-%m-%dT%H:%M:%S")
-        response = await client.patch(
-            f"/v1/users/license-keys/{lk.id}",
-            json={
-                "usage": 4,
-                "limit_usage": 10,
-                "limit_activations": 5,
-                "expires_at": expires_at,
-            },
-        )
-        assert response.status_code == 200
-        updated = response.json()
-        assert updated["usage"] == 4
-        assert updated["limit_usage"] == 10
-        assert updated["limit_activations"] == 5
-        assert updated["expires_at"] == expires_at
-
-
-@pytest.mark.asyncio
-@pytest.mark.http_auto_expunge
-class TestLicenseKeyActionEndpoints:
+class TestUserLicenseKeyEndpoints:
 
     @pytest.mark.auth(
         AuthSubjectFixture(subject="user"),
@@ -618,7 +491,7 @@ class TestLicenseKeyActionEndpoints:
         )
         assert response.status_code == 200
 
-        response = await client.get(f"/v1/users/license-keys/{lk.id}")
+        response = await client.get(f"/v1/license-keys/{lk.id}")
         assert response.status_code == 200
         fetched = response.json()
         assert len(fetched["activations"]) == 1
@@ -656,7 +529,7 @@ class TestLicenseKeyActionEndpoints:
         activated = response.json()
         new_activation_id = activated["id"]
 
-        response = await client.get(f"/v1/users/license-keys/{lk.id}")
+        response = await client.get(f"/v1/license-keys/{lk.id}")
         assert response.status_code == 200
         fetched = response.json()
         assert len(fetched["activations"]) == 1
